refactor: 更新演示项目中SDK的使用方法
