--- carbon-client-cli/pom.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2018-2023. Chengdu WeiSiFan Technology Co., Ltd.
-  ~ Carbon Integration SDK is licensed under Mulan PSL v2.
-  ~
-  ~ You can use this software according to the terms and conditions of the Mulan PSL v2.
-  ~ You may obtain a copy of Mulan PSL v2 at: http://license.coscl.org.cn/MulanPSL2
-  ~
-  ~ THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
-  ~ EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
-  ~ MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
-  ~ See the Mulan PSL v2 for more details.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.zhibaocloud</groupId>
-    <artifactId>carbon-integration</artifactId>
-    <version>0.1.0</version>
-  </parent>
-
-  <artifactId>carbon-client-cli</artifactId>
-  <name>Carbon Client Sample Application</name>
-  <description>客户端数据推送示例应用程序</description>
-
-  <properties>
-    <maven.deploy.skip>true</maven.deploy.skip>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.projectlombok</groupId>
-      <artifactId>lombok</artifactId>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.github.jsonzou</groupId>
-      <artifactId>jmockdata</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-demo-secret</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-logging</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-client-spring-boot-starter</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-  
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-maven-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-</project>
