--- carbon-client-cli/src/main/resources/application.yaml
@@ -1,3 +1,0 @@
-spring:
-  application:
-    name: carbon-client-cli
