+++ carbon-client-spring-boot-starter/pom.xml
@@ -20,6 +20,7 @@
     <groupId>com.zhibaocloud</groupId>
     <artifactId>carbon-integration</artifactId>
     <version>0.1.0</version>
+    <relativePath>../pom.xml</relativePath>
   </parent>
 
   <artifactId>carbon-client-spring-boot-starter</artifactId>
