--- carbon-demo-secret/pom.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2018-2023. Chengdu WeiSiFan Technology Co., Ltd.
-  ~ Carbon Integration SDK is licensed under Mulan PSL v2.
-  ~
-  ~ You can use this software according to the terms and conditions of the Mulan PSL v2.
-  ~ You may obtain a copy of Mulan PSL v2 at: http://license.coscl.org.cn/MulanPSL2
-  ~
-  ~ THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
-  ~ EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
-  ~ MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
-  ~ See the Mulan PSL v2 for more details.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.zhibaocloud</groupId>
-    <artifactId>carbon-integration</artifactId>
-    <version>0.1.0</version>
-  </parent>
-
-  <artifactId>carbon-demo-secret</artifactId>
-  <name>Carbon Demo Shared Secret</name>
-  <description>工具模块，确保客户端和服务端使用相同的密钥</description>
-
-  <properties>
-    <maven.deploy.skip>true</maven.deploy.skip>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-exchange</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-</project>
