+++ carbon-server-demo/pom.xml
@@ -23,9 +23,10 @@
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <artifactId>carbon-coverage</artifactId>
-  <name>Carbon Test Coverage Report</name>
-  <description>测试覆盖率报告</description>
+  <artifactId>carbon-server-demo</artifactId>
+  <name>Carbon Demo Server</name>
+  <description>Demo 服务，担当数据推送的接收方，验证数据的有效性，以及演示会如何接收推送数据
+  </description>
 
   <properties>
     <maven.deploy.skip>true</maven.deploy.skip>
@@ -33,33 +34,55 @@
 
   <dependencies>
     <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
       <groupId>com.zhibaocloud</groupId>
       <artifactId>carbon-exchange</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-client-sdk</artifactId>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <scope>runtime</scope>
     </dependency>
     <dependency>
       <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-client-spring-boot-starter</artifactId>
+      <artifactId>carbon-server-sdk</artifactId>
     </dependency>
     <dependency>
       <groupId>com.zhibaocloud</groupId>
       <artifactId>carbon-server-spring-boot-starter</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-server-demo</artifactId>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-web</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.zhibaocloud</groupId>
-      <artifactId>carbon-client-demo</artifactId>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-logging</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.tngtech.archunit</groupId>
-      <artifactId>archunit-junit5</artifactId>
-      <version>1.0.1</version>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-data-jpa</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-validation</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springdoc</groupId>
+      <artifactId>springdoc-openapi-webmvc-core</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-test</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.github.jsonzou</groupId>
+      <artifactId>jmockdata</artifactId>
       <scope>test</scope>
     </dependency>
   </dependencies>
@@ -67,17 +90,8 @@
   <build>
     <plugins>
       <plugin>
-        <groupId>org.jacoco</groupId>
-        <artifactId>jacoco-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>report-aggregate</id>
-            <phase>verify</phase>
-            <goals>
-              <goal>report-aggregate</goal>
-            </goals>
-          </execution>
-        </executions>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-maven-plugin</artifactId>
       </plugin>
     </plugins>
   </build>
