+++ carbon-server-demo/src/main/resources/application.yaml
