+++ pom.xml
@@ -54,11 +54,12 @@
     <module>carbon-exchange</module>
     <module>carbon-client-sdk</module>
     <module>carbon-client-spring-boot-starter</module>
-    <module>carbon-client-cli</module>
-    <module>carbon-mock-server</module>
-    <module>carbon-coverage</module>
-    <module>carbon-demo-secret</module>
+    <module>carbon-server-sdk</module>
     <module>carbon-server-spring-boot-starter</module>
+
+    <module>carbon-server-demo</module>
+    <module>carbon-client-demo</module>
+    <module>carbon-coverage</module>
   </modules>
 
   <properties>
@@ -158,12 +159,17 @@
       </dependency>
       <dependency>
         <groupId>com.zhibaocloud</groupId>
+        <artifactId>carbon-server-sdk</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.zhibaocloud</groupId>
         <artifactId>carbon-client-spring-boot-starter</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
         <groupId>com.zhibaocloud</groupId>
-        <artifactId>carbon-client-cli</artifactId>
+        <artifactId>carbon-client-demo</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -173,12 +179,7 @@
       </dependency>
       <dependency>
         <groupId>com.zhibaocloud</groupId>
-        <artifactId>carbon-mock-server</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>com.zhibaocloud</groupId>
-        <artifactId>carbon-demo-secret</artifactId>
+        <artifactId>carbon-server-demo</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
